rename current switch -> tabs
