+++ frontend/components/tabs/index.ts
