+++ frontend/components/tabs/types.d.ts
@@ -1,0 +1,8 @@
+import type { OptionType } from 'types';
+
+export interface TabsProps {
+  options: OptionType[];
+  selectedValue: string;
+  className?: string;
+  onChange?: (selectedValue: string) => void;
+}
